refactor(workers): restructured workers, seperating actual workers from spawners
