+++ workers/workers/stlStreamWorker.src.js
@@ -1,5 +1,5 @@
 const ParentStream = require('workerstream/parent')
-import { default as makeStlStreamParser } from '../parsers/stl/parseStream'
+import { default as makeStlStreamParser } from '../../parsers/stl/parseStream'
 
 module.exports = function () {
   console.log('here in worker')
