+++ workers/workers/worker.src.js
@@ -1,8 +1,14 @@
-import parseSteps from '../parsers/stl/parseHelpers.js'
+/*function parseSteps (data) {
+  return {
+    positions: new ArrayBuffer(1),//data.size), //[0, 1, 2],
+    normals: new ArrayBuffer(2)//data.size),//[2, 1, 0]
+  }
+}*/
+import parseSteps from '../../parsers/stl/parseHelpers.js'
 
 self.onmessage = function (event) {
   //console.log('event',event.data)
-  let result = parseSteps(event.data.buffer)
+  const result = parseSteps(event.data)
 
   let positions = result.positions.buffer
   let normals = result.normals.buffer
@@ -14,6 +20,4 @@ self.onmessage = function (event) {
   if ('close' in self) {
     self.close()
   }
-
-
 }
