+++ workers/workers/worker2.src.js
@@ -1,14 +1,8 @@
-/*function parseSteps (data) {
-  return {
-    positions: new ArrayBuffer(1),//data.size), //[0, 1, 2],
-    normals: new ArrayBuffer(2)//data.size),//[2, 1, 0]
-  }
-}*/
-import parseSteps from '../parsers/stl/parseHelpers.js'
+import parseSteps from '../../parsers/stl/parseHelpers.js'
 
 self.onmessage = function (event) {
   //console.log('event',event.data)
-  const result = parseSteps(event.data)
+  let result = parseSteps(event.data.buffer)
 
   let positions = result.positions.buffer
   let normals = result.normals.buffer
@@ -20,4 +14,6 @@ self.onmessage = function (event) {
   if ('close' in self) {
     self.close()
   }
+
+
 }
